# Please enter a commit message to explain why this merge is necessary,
# especially if it merges an updated upstream into a topic branch.
#
# Lines starting with '#' will be ignored, and an empty message aborts
# the commit.


Former-commit-id: b325a7b9c4a594360244e6c221fc5e2c3c172789
Former-commit-id: 50c8b81c58f76059035a94c6642ad6f820119052

+++ kernel/kernel-masterdata-service/src/main/java/io/mosip/kernel/masterdata/service/impl/LocationServiceImpl.java
@@ -388,9 +388,9 @@ public class LocationServiceImpl implements LocationService {
 	}
 
 	@Override
-	public Map<Integer, List<Location>> getLocationByLangCodeAndHierarchyLevel(String langCode,
-			Integer hierarchyLevel) {
-		Map<Integer, List<Location>> map = new TreeMap<>();
+	public Map<Short, List<Location>> getLocationByLangCodeAndHierarchyLevel(String langCode,
+			Short hierarchyLevel) {
+		Map<Short, List<Location>> map = new TreeMap<>();
 		List<Location> locations = locationRepository.getAllLocationsByLangCodeAndLevel(langCode, hierarchyLevel);
 		if (!EmptyCheckUtils.isNullEmpty(locations)) {
 			for (Location location : locations) {
